SECURITY.md: add go-ports specific caution
Our motto in go-ports should be "first, do no harm" so let's go out of our way to be sure not to generate spurious trouble reports to the Go team.

Even better would be a dedicated go-ports security response, but only if it can be assured not to delay any reports that apply also to upstream. I don't know how we can do that yet.