runtime: fix MMU initialization for tamago/arm
The MMU initialization was not done correctly leaving the bootloader
defined Translation Table on the real hardware and leaving the MMU
disabled in qemu.

It was witnessed that on the ARM architecture ARMv6 included in Broadcom
BCM2835 the LDREX instructions triggers an exception if the MMU is not
enabled.

Therefore we invoke mmuinit() before any LDREX instruction.